remove more deps to ui
