+++ ui/components/application/OrganisationForm.tsx
@@ -11,11 +11,8 @@ import {
   OptionType,
 } from "common/types/common";
 import { breakpoints } from "common/src/common/style";
-import {
-  CheckboxWrapper,
-  TwoColumnContainer,
-  FormSubHeading,
-} from "../common/common";
+import { CheckboxWrapper } from "common/src/reservation-form/components";
+import { TwoColumnContainer, FormSubHeading } from "../common/common";
 import EmailInput from "./EmailInput";
 import BillingAddress from "./BillingAddress";
 import Buttons from "./Buttons";
