+++ ui/components/common/common.ts
@@ -67,13 +67,3 @@ export const CenterSpinner = styled(LoadingSpinner).attrs({
 })`
   margin: 0 auto var(--spacing-2-xl) auto;
 `;
-
-export const CheckboxWrapper = styled.div<{ $break?: boolean }>`
-  ${({ $break }) => $break && "grid-column: 1 / -1"};
-  margin-top: 2.5em;
-  margin-bottom: auto;
-
-  @media (max-width: ${breakpoints.m}) {
-    margin-top: 0;
-  }
-`;
