+++ ui/modules/types.ts
@@ -1,73 +1,3 @@
-export type Inputs = {
-  pk: number;
-  reserveeFirstName: string;
-  reserveeLastName: string;
-  reserveePhone: string;
-  name: string;
-  description: string;
-  spaceTerms: boolean;
-  resourceTerms: boolean;
-  purpose: number;
-  numPersons: number;
-  ageGroup: number;
-  reserveeAddressStreet: string;
-  reserveeAddressZip: string;
-  reserveeAddressCity: string;
-  reserveeEmail: string;
-  reserveeOrganisationName: string;
-  showBillingAddress?: boolean;
-  billingFirstName: string;
-  billingLastName: string;
-  billingPhone: string;
-  billingEmail: string;
-  billingAddressStreet: string;
-  billingAddressCity: string;
-  billingAddressZip: string;
-  homeCity: number;
-  applyingForFreeOfCharge: boolean;
-  freeOfChargeReason: string;
-  reserveeIsUnregisteredAssociation?: boolean;
-  reserveeId?: number;
-};
-
-export type Reservation = {
-  pk: number;
-  begin: string;
-  end: string;
-  reservationUnitPks: number[];
-  reserveeFirstName?: string;
-  reserveeLastName?: string;
-  reserveePhone?: string;
-  name?: string;
-  user?: { email: string };
-  description?: string;
-  calendarUrl?: string;
-  state?: string;
-  price?: number;
-  taxPercentageValue?: number;
-  spaceTerms?: boolean;
-  resourceTerms?: boolean;
-  purpose?: number;
-  numPersons?: number;
-  ageGroup?: number;
-  reserveeAddressStreet?: string;
-  reserveeAddressZip?: string;
-  reserveeAddressCity?: string;
-  reserveeEmail?: string;
-  reserveeOrganisationName?: string;
-  showBillingAddress?: boolean;
-  billingFirstName?: string;
-  billingLastName?: string;
-  billingPhone?: string;
-  billingEmail?: string;
-  billingAddressStreet?: string;
-  billingAddressCity?: string;
-  billingAddressZip?: string;
-  homeCity?: number;
-  applyingForFreeOfCharge?: boolean;
-  freeOfChargeReason?: string;
-};
-
 export type ReservationStep = {
   label: string;
   state: 0 | 1 | 2;
